Added some state change logging to the Pony class, available in Debug mode.
Changed the parameter in the "standard" call to Paint which should allow the override to kick in correctly when destination is reached.
Changed SelectBehavior to default to the first behavior if one could not be selected at random and the pony was interacting, this should prevent a pony getting stuck replaying the interaction behavior after its intended duration.