Fixed error in NewPonyDialog where directory for new pony did need to be created early in the process. If the dialog is canceled, any directory created is removed.
Improved robustness of invalid character checks when creating a new pony name.
Fixed an error in the saving routine for the editor that would crash if the animator wasn't running (a pony wasn't yet selected, but a new one was created)
Fixed a bad assumption about File.Replace happily replacing a non-existent file. It requires the destination to exist.
Improved automatic sizing of the exception dialog to better handle long messages, and not size larger than the screen if possible.