Added ItemsViewerBase - for the item collections in the new editor. A simple initial BehaviorsViewer has been implemented.
Added several new buttons and context menus to the new editor. These will allow access to the preview, and the new items lists.
Changed PonyPreview to refresh bounds when the area has changed, rather than just on every update. It will also set DisplayBounds on the WinFormSpriteInterface, which gives a decent impression that the interface is actually part of the editor since it is nicely clipped at the edges.
Changed StringCollectionParser to not SkipParse when an earlier failure occurred. This may create spurious errors, but ensures if a single fatal problem occurs, that later valid items are still parsed. This helps the editor windows which may otherwise update with fallback values when a fatal error occurs, and then push the fallback values to the ini string if the problem is fixed via the UI - thus the wiping the original values.